Abstract ImperativeMessagesRepository api and ImperativeMessagesService
Introduce type Id, that allows us to treat any imperative api like an
api in some F[_]. That way, we can abstract the api of the repository
and the full implementation of the service. We now have only one
implementation for the service that can be used for imperative, Result
and Async programs.