Make parameters of type CanContinue implicit
Looking at Application.scala we can see that CanContinue instances are
tied to the type they work for. In fact, it seems natural that given
a type for F[_] there should be one canonical way to implement
CanContinue for that type. Making CanContinue parameters implicit
implies that we want this canonical instance.

CanContinue will become what we call a typeclass. CanContinue is a class
of types in the sense that types with a canonical CanContinue instance
all behave in a similar way. This allows us to add any given type to
the type class by implementing an instance of CanContinue for it, like
we did for Id. This is something we could not do if we had choosen to
abstract the common behaviour of Async and Result with inheritance (by
making both implement a common parent trait).